search function and cofirm delete popup

+++ client/src/app/dashboard/list/assignments/page.tsx
@@ -5,7 +5,7 @@ import Image from "next/image";
 import { FaPlus } from "react-icons/fa";
 
 import Table from "@/app/component/Table";
-import TableSearch from "@/app/component/TableSearch";
+import ResultTableSearch from "@/app/component/ResultTableSearch";
 import Pagination from "@/app/component/Pagination";
 import AssignmentForm from "@/app/component/forms/AssignmentForm";
 
@@ -24,6 +24,7 @@ const columns = [
   { header: "Teacher", accessor: "teacherName", className: "hidden md:table-cell" },
   { header: "Due Date", accessor: "dueDate", className: "hidden md:table-cell" },
   { header: "Actions", accessor: "action" },
+  
 ];
 
 const AssignmentListPage = () => {
@@ -31,6 +32,8 @@ const AssignmentListPage = () => {
   const [isFormOpen, setIsFormOpen] = useState(false);
   const [isEditMode, setIsEditMode] = useState(false);
   const [editItem, setEditItem] = useState<Assignment | null>(null);
+  const [searchTerm, setSearchTerm] = useState("");
+
 
   useEffect(() => {
     fetchAssignments();
@@ -78,6 +81,14 @@ const AssignmentListPage = () => {
   };
 
   const handleDeleteAssignment = async (id: string) => {
+
+    const confirmDelete = window.confirm(
+      "Are you sure you want to delete this result?"
+    );
+    if (!confirmDelete) {
+      return;
+    }
+    
     try {
       const res = await fetch(`http://localhost:5000/api/assignments/${id}`, {
         method: "DELETE",
@@ -111,6 +122,10 @@ const AssignmentListPage = () => {
     setIsFormOpen(false);
   };
 
+  const filteredAssignment = assignments.filter((item) =>
+    item.subjectName.toLowerCase().includes(searchTerm.toLowerCase())
+  );
+
   const renderRow = (item: Assignment) => (
     <tr key={item._id} className="border-b border-gray-200 even:bg-slate-50 text-sm hover:bg-cbPurpleLight">
       <td className="p-4">{item.subjectName}</td>
@@ -135,7 +150,10 @@ const AssignmentListPage = () => {
       <div className="flex items-center justify-between">
         <h1 className="hidden md:block text-lg font-semibold">All Assignments</h1>
         <div className="flex flex-col md:flex-row items-center gap-4 w-full md:w-auto">
-          <TableSearch />
+        <ResultTableSearch
+            value={searchTerm}
+            onChange={(e) => setSearchTerm(e.target.value)}
+          />
           <div className="flex items-center gap-4 self-end">
             <button className="w-8 h-8 flex items-center justify-center rounded-full bg-cbYellow">
               <Image src="/filter.png" alt="Filter" width={14} height={14} />
@@ -150,7 +168,7 @@ const AssignmentListPage = () => {
         </div>
       </div>
 
-      <Table columns={columns} renderRow={renderRow} data={assignments} />
+      <Table columns={columns} renderRow={renderRow} data={filteredAssignment} />
       <Pagination />
 
       {isFormOpen && (
