+++ client/src/app/dashboard/list/exams/page.tsx
@@ -4,7 +4,7 @@ import { useEffect, useState } from "react";
 import Image from "next/image";
 import { FaPlus } from "react-icons/fa";
 import ExamForm from "@/app/component/forms/ExamForm";
-import TableSearch from "@/app/component/TableSearch";
+import ResultTableSearch from "@/app/component/ResultTableSearch";
 import Table from "@/app/component/Table";
 import Pagination from "@/app/component/Pagination";
 
@@ -31,6 +31,7 @@ const ExamListPage = () => {
   const [isFormOpen, setIsFormOpen] = useState(false);
   const [isEditMode, setIsEditMode] = useState(false);
   const [editItem, setEditItem] = useState<Exam | null>(null);
+  const [searchTerm, setSearchTerm] = useState("");
 
   useEffect(() => {
     fetchExams();
@@ -76,6 +77,14 @@ const ExamListPage = () => {
   };
 
   const handleDeleteExam = async (id: string) => {
+
+    const confirmDelete = window.confirm(
+      "Are you sure you want to delete this result?"
+    );
+    if (!confirmDelete) {
+      return;
+    }
+    
     try {
       const res = await fetch(`http://localhost:5000/api/exams/${id}`, {
         method: "DELETE",
@@ -109,6 +118,12 @@ const ExamListPage = () => {
     setIsFormOpen(false);
   };
 
+  const filteredResults = exams.filter((item) =>
+    item.subject.toLowerCase().includes(searchTerm.toLowerCase())
+  );
+
+  
+
   const renderRow = (item: Exam) => (
     <tr key={item._id} className="border-b border-gray-200 even:bg-slate-50 text-sm hover:bg-cbPurpleLight">
       <td className="p-4">{item.subject}</td>
@@ -133,7 +148,10 @@ const ExamListPage = () => {
       <div className="flex items-center justify-between">
         <h1 className="hidden md:block text-lg font-semibold">All Exams</h1>
         <div className="flex flex-col md:flex-row items-center gap-4 w-full md:w-auto">
-          <TableSearch />
+        <ResultTableSearch
+            value={searchTerm}
+            onChange={(e) => setSearchTerm(e.target.value)}
+          />
           <div className="flex items-center gap-4 self-end">
             <button className="w-8 h-8 flex items-center justify-center rounded-full bg-cbYellow">
               <Image src="/filter.png" alt="Filter" width={14} height={14} />
@@ -150,7 +168,7 @@ const ExamListPage = () => {
         </div>
       </div>
 
-      <Table columns={columns} renderRow={renderRow} data={exams} />
+      <Table columns={columns} renderRow={renderRow} data={filteredResults} />
       <Pagination />
 
       {isFormOpen && (
